Security hardening from crypto audit
- Add memory zeroization for secrets (zeroBytes helper, defer cleanup)
- Add KeyDerivation.Zero() method for secure cleanup
- Fix MaxPriceValue: reduce from 1e12 to uint32 max (4.3B) to prevent overflow
- Add private key validation (length, curve order bounds)
- Add panic recovery middleware for all HTTP handlers
- Add rate limiting to /api/quote endpoint (was only on webhook)
- Fix race condition in webhook handler (copy channel ref under lock)
- Add panicsRecovered Prometheus metric

+++ shared/types.go
@@ -120,8 +120,8 @@ func (r *HttpRequestData) Validate() error {
 		if price <= 0 {
 			return fmt.Errorf("threshold price must be positive, got %.2f", price)
 		}
-		if price > MaxPriceValue {
-			return fmt.Errorf("threshold price exceeds maximum %.0f, got %.2f", MaxPriceValue, price)
+		if price > float64(MaxPriceValue) {
+			return fmt.Errorf("threshold price exceeds maximum %d, got %.2f", MaxPriceValue, price)
 		}
 	}
